+++ mcumgr-toolkit/src/client.rs
@@ -32,6 +32,12 @@ use crate::{
 /// Matches Zephyr default value of [MCUMGR_TRANSPORT_NETBUF_SIZE](https://github.com/zephyrproject-rtos/zephyr/blob/v4.2.1/subsys/mgmt/mcumgr/transport/Kconfig#L40).
 const ZEPHYR_DEFAULT_SMP_FRAME_SIZE: usize = 384;
 
+/// The default timeout used by this crate
+pub const DEFAULT_TIMEOUT_MS: u64 = 500;
+
+/// THe default retry count used by this crate
+pub const DEFAULT_RETRIES: u8 = 5;
+
 /// A high-level client for Zephyr's MCUmgr SMP protocol.
 ///
 /// This struct is the central entry point of this crate.
@@ -220,7 +226,6 @@ impl MCUmgrClient {
     /// # use mcumgr_toolkit::MCUmgrClient;
     /// # fn main() {
     /// let serial = serialport::new("COM42", 115200)
-    ///     .timeout(std::time::Duration::from_millis(500))
     ///     .open()
     ///     .unwrap();
     ///
@@ -231,7 +236,7 @@ impl MCUmgrClient {
         serial: T,
     ) -> Self {
         Self {
-            connection: Connection::new(SerialTransport::new(serial), 5),
+            connection: Connection::new(SerialTransport::new(serial), DEFAULT_RETRIES),
             smp_frame_size: ZEPHYR_DEFAULT_SMP_FRAME_SIZE.into(),
         }
     }
@@ -369,6 +374,14 @@ impl MCUmgrClient {
         self.connection
             .set_timeout(timeout)
             .map_err(MCUmgrClientError::SetTimeoutFailed)
+    }
+
+    /// Changes the retry amount.
+    ///
+    /// When the device encounters a transport error, it will retry
+    /// this many times until giving up.
+    pub fn set_retries(&self, retries: u8) {
+        self.connection.set_retries(retries)
     }
 
     /// Checks if the device is alive and responding.